rankspace

Enum RankSpaceError

pub enum RankSpaceError {
    DimMismatch {
        dims: usize,
        strides: usize,
    },
    DuplicateDim {
        name: String,
    },
    UnknownDim {
        name: String,
    },
    DimIndexOutOfRange {
        dim: String,
        index: usize,
        size: usize,
    },
    NonrectangularStrides,
    OverlappingStrides,
    StrideTooSmall {
        stride: usize,
        space: usize,
    },
    RankArithmeticOverflow,
    IncompatibleEmbedding,
    ZeroStep,
}

Errors produced by rank-space construction and operations.

Variants

DimMismatch

The number of dimensions and strides differ.

Fields

dims: usize

strides: usize

DuplicateDim

The same dimension name was provided more than once.

Fields

name: String

UnknownDim

The requested dimension does not exist.

Fields

name: String

DimIndexOutOfRange

The requested dimension index is out of bounds.

Fields

dim: String

index: usize

size: usize

NonrectangularStrides

The affine strides do not define a rectangular rank space.

OverlappingStrides

Two non-unit dimensions use the same stride.

StrideTooSmall

A stride overlaps the previous coordinate space.

Fields

stride: usize

space: usize

RankArithmeticOverflow

Rank arithmetic overflowed.

IncompatibleEmbedding

The requested embedding cannot be represented as one rank rectangle.

ZeroStep

Dimension ranges must have nonzero steps.